backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
icalendar==5.0.11
python-dateutil==2.8.2
pydantic==2.6.1
python-dotenv==1.0.1
pytz==2024.1
openai==1.66.3
python-multipart==0.0.9
starlette==0.36.3
